Fix a simple error in the sampling algorithm
A test that catches this is a statistical test. Distributions.jl has tests like this. I could implement a similar test, but I think it would be better to define my ziggurat samplers as a subtype of Distributions.Sampleable{Univariate, Continuous}. That change is forthcoming soon. Then, I will be able to easily reuse the Distributions test code.